Extract shared code into common library
